+++ bika/lims/browser/analyses/view.py
@@ -927,7 +927,8 @@ class AnalysesView(BikaListingView):
         max_str = results_range.get('max', '')
         min_str = api.is_floatable(min_str) and "{0}".format(min_str) or ""
         max_str = api.is_floatable(max_str) and "{0}".format(max_str) or ""
-        specs = ", ".join([val for val in [min_str, max_str] if val])
+        # Join with semi-colon to avoid confusion with commas as decimal mark
+        specs = "; ".join([val for val in [min_str, max_str] if val])
         if not specs:
             return
         item["Specification"] = "[{}]".format(specs)