+++ bika/lims/controlpanel/bika_instruments.py
@@ -5,144 +5,194 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
-from AccessControl import ClassSecurityInfo
-from Products.ATContentTypes.content import schemata
-from Products.Archetypes import atapi
-from Products.Archetypes.ArchetypeTool import registerType
-from Products.CMFCore import permissions
-from Products.CMFCore.utils import getToolByName
-from bika.lims.browser import BrowserView
+import collections
+
+from bika.lims import bikaMessageFactory as _
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.config import PROJECTNAME
-from bika.lims import bikaMessageFactory as _
-from bika.lims.utils import t
-from bika.lims.content.bikaschema import BikaFolderSchema
 from bika.lims.interfaces import IInstruments
-from plone.app.layout.globals.interfaces import IViewView
 from plone.app.content.browser.interfaces import IFolderContentsView
-from plone.app.folder.folder import ATFolder, ATFolderSchema
+from plone.app.folder.folder import ATFolder
+from plone.app.folder.folder import ATFolderSchema
+from plone.app.layout.globals.interfaces import IViewView
+from bika.lims.utils import get_link
+from Products.Archetypes import atapi
+from Products.ATContentTypes.content import schemata
 from zope.interface.declarations import implements
-from operator import itemgetter
+
+
+# TODO: Separate content and view into own modules!
+
 
 class InstrumentsView(BikaListingView):
     implements(IFolderContentsView, IViewView)
+
     def __init__(self, context, request):
         super(InstrumentsView, self).__init__(context, request)
-        self.catalog = 'bika_setup_catalog'
-        self.contentFilter = {'portal_type': 'Instrument',
-                              'sort_on': 'sortable_title'}
-        self.context_actions = {_('Add'):
-                                {'url': 'createObject?type_name=Instrument',
-                                 'icon': '++resource++bika.lims.images/add.png'}}
+
+        self.catalog = "bika_setup_catalog"
+        self.contentFilter = {
+            "portal_type": "Instrument",
+            "sort_on": "sortable_title",
+            "sort_order": "ascending",
+        }
+
+        self.context_actions = {
+            _("Add"): {
+                "url": "createObject?type_name=Instrument",
+                "permisison": "Add portal content",
+                "icon": "++resource++bika.lims.images/add.png"}
+        }
+
         self.title = self.context.translate(_("Instruments"))
-        self.icon = self.portal_url + "/++resource++bika.lims.images/instrument_big.png"
         self.description = ""
+        self.icon = "{}/{}".format(
+            self.portal_url,
+            "/++resource++bika.lims.images/instrument_big.png"
+        )
+
         self.show_sort_column = False
         self.show_select_row = False
         self.show_select_column = True
         self.pagesize = 25
 
-        self.columns = {
-            'Title': {'title': _('Instrument'),
-                      'index': 'sortable_title'},
-            'Type': {'title': _('Type'),
-                     'index': 'getInstrumentTypeName',
-                     'toggle': True,
-                     'sortable': True},
-            'Brand': {'title': _('Brand'),
-                      'toggle': True},
-            'Model': {'title': _('Model'),
-                      'index': 'getModel',
-                      'toggle': True},
-            'ExpiryDate': {'title': _('Expiry Date'),
-                           'toggle': True},
-            'WeeksToExpire': {'title': _('Weeks To Expire'),
-                           'toggle': False},
-            'Methods': {'title': _('Methods'),
-                           'toggle': True},
-            }
+        self.columns = collections.OrderedDict((
+            ("Title", {
+                "title": _("Instrument"),
+                "index": "sortable_title"}),
+            ("Type", {
+                "title": _("Type"),
+                "index": "getInstrumentTypeName",
+                "toggle": True,
+                "sortable": True}),
+            ("Brand", {
+                "title": _("Brand"),
+                "sortable": False,
+                "toggle": True}),
+            ("Model", {
+                "title": _("Model"),
+                "index": "getModel",
+                "toggle": True}),
+            ("ExpiryDate", {
+                "title": _("Expiry Date"),
+                "sortable": False,
+                "toggle": True}),
+            ("WeeksToExpire", {
+                "title": _("Weeks To Expire"),
+                "sortable": False,
+                "toggle": False}),
+            ("Methods", {
+                "title": _("Methods"),
+                "sortable": False,
+                "toggle": True}),
+        ))
 
         self.review_states = [
-            {'id':'default',
-             'title': _('Active'),
-             'contentFilter': {'inactive_state': 'active'},
-             'transitions': [{'id':'deactivate'}, ],
-             'columns': ['Title',
-                         'Type',
-                         'Brand',
-                         'Model',
-                         'ExpiryDate',
-                         'WeeksToExpire',
-                         'Methods']},
-            {'id':'inactive',
-             'title': _('Dormant'),
-             'contentFilter': {'inactive_state': 'inactive'},
-             'transitions': [{'id':'activate'}, ],
-             'columns': ['Title',
-                         'Type',
-                         'Brand',
-                         'Model',
-                         'ExpiryDate',
-                         'WeeksToExpire',
-                         'Methods']},
-            {'id':'all',
-             'title': _('All'),
-             'contentFilter':{},
-             'columns': ['Title',
-                         'Type',
-                         'Brand',
-                         'Model',
-                         'ExpiryDate',
-                         'WeeksToExpire',
-                         'Methods']},
-            ]
+            {
+                "id": "default",
+                "title": _("Active"),
+                "contentFilter": {"inactive_state": "active"},
+                "transitions": [{"id": "deactivate"}, ],
+                "columns": self.columns.keys(),
+            }, {
+                "id": "inactive",
+                "title": _("Dormant"),
+                "contentFilter": {"inactive_state": "inactive"},
+                "transitions": [{"id": "activate"}, ],
+                "columns": self.columns.keys(),
+            }, {
+                "id": "all",
+                "title": _("All"),
+                "contentFilter": {},
+                "columns": self.columns.keys(),
+            },
+        ]
 
-    def folderitems(self):
-        items = BikaListingView.folderitems(self)
-        for x in range(len(items)):
-            if not items[x].has_key('obj'): continue
-            obj = items[x]['obj']
+    def before_render(self):
+        """Before template render hook
+        """
+        # Don't allow any context actions on the Instruments folder
+        self.request.set("disable_border", 1)
 
-            itype = obj.getInstrumentType()
-            items[x]['Type'] = itype.Title() if itype else ''
-            ibrand = obj.getManufacturer()
-            items[x]['Brand'] = ibrand.Title() if ibrand else ''
-            items[x]['Model'] = obj.getModel()
+    def folderitem(self, obj, item, index):
+        """Service triggered each time an item is iterated in folderitems.
 
-            data = obj.getCertificateExpireDate()
-            if data is None:
-                items[x]['ExpiryDate'] = _("No date set")
-            else:
-                items[x]['ExpiryDate'] = data.asdatetime().strftime(self.date_format_short)
+        The use of this service prevents the extra-loops in child objects.
 
-            if obj.isOutOfDate():
-                items[x]['WeeksToExpire'] = _("Out of date")
-            else:
-                weeks, days = obj.getWeeksToExpire()
-                weeks_to_expire = _("{} weeks and {} day(s)".format(str(weeks), str(days)))
-                items[x]['WeeksToExpire'] = weeks_to_expire
+        :obj: the instance of the class to be foldered
+        :item: dict containing the properties of the object to be used by
+            the template
+        :index: current index of the item
+        """
 
-            methods = obj.getMethods()
-            urls = []
-            titles = []
-            for method in methods:
-                url = method.absolute_url()
-                title = method.Title()
-                titles.append(title)
-                urls.append("<a href='{0}'>{1}</a>".format(url, title))
+        title = obj.Title()
+        url = obj.absolute_url()
 
-            items[x]["Methods"] = ", ".join(titles)
-            items[x]["replace"]["Methods"] = ", ".join(urls)
-            items[x]["replace"]["Title"] = "<a href='{0}'>{1}</a>".format(
-                obj.absolute_url(), obj.Title())
+        item["Title"] = title
+        item["replace"]["Title"] = get_link(url, value=title)
 
-        return items
+        instrument_type = obj.getInstrumentType()
+        if instrument_type:
+            url = instrument_type.absolute_url()
+            title = instrument_type.Title()
+            item["Type"] = instrument_type.Title()
+            item["replace"]["Type"] = get_link(url, value=title)
+        else:
+            item["Type"] = ""
+
+        instrument_brand = obj.getManufacturer()
+        if instrument_brand:
+            url = instrument_brand.absolute_url()
+            title = instrument_brand.Title()
+            item["Brand"] = instrument_brand.Title()
+            item["replace"]["Brand"] = get_link(url, value=title)
+        else:
+            item["Brand"] = ""
+
+        instrument_model = obj.getModel()
+        if instrument_model:
+            item["Model"] = instrument_model
+        else:
+            item["Model"] = ""
+
+        expiry_date = obj.getCertificateExpireDate()
+        if expiry_date is None:
+            item["ExpiryDate"] = _("No date set")
+        else:
+
+            item["ExpiryDate"] = expiry_date.asdatetime().strftime(
+                self.date_format_short)
+
+        if obj.isOutOfDate():
+            item["WeeksToExpire"] = _("Out of date")
+        else:
+            weeks, days = obj.getWeeksToExpire()
+            weeks_to_expire = _("{} weeks and {} day(s)".format(
+                str(weeks), str(days)))
+            item['WeeksToExpire'] = weeks_to_expire
+
+        methods = obj.getMethods()
+        if methods:
+            links = map(
+                lambda m: get_link(m.absolute_url(),
+                                   value=m.Title(),
+                                   css_class="link"),
+                methods)
+            item["replace"]["Methods"] = ", ".join(links)
+
+        return item
+
 
 schema = ATFolderSchema.copy()
+
+
 class Instruments(ATFolder):
+    """Instruments Folder
+    """
     implements(IInstruments)
     displayContentsTab = False
     schema = schema
 
-schemata.finalizeATCTSchema(schema, folderish = True, moveDiscussion = False)
+
+schemata.finalizeATCTSchema(schema, folderish=True, moveDiscussion=False)
 atapi.registerType(Instruments, PROJECTNAME)